Chapter 7 Renaming/Moving API classes to data.models package. Using Moshi's codegen.
