+++ 07-networking/projects/final/app/src/androidTest/java/com/kodeco/recipefinder/ExampleInstrumentedTest.kt
@@ -34,13 +34,11 @@
 
 package com.kodeco.recipefinder
 
-import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
-
+import androidx.test.platform.app.InstrumentationRegistry
+import org.junit.Assert.*
 import org.junit.Test
 import org.junit.runner.RunWith
-
-import org.junit.Assert.*
 
 /**
  * Instrumented test, which will execute on an Android device.
