+++ 07-networking/projects/final/app/src/main/java/com/kodeco/recipefinder/data/models/Recipe.kt
@@ -34,6 +34,9 @@
 
 package com.kodeco.recipefinder.data.models
 
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
 data class Recipe(
   val id: Int,
   val title: String,
