+++ 07-networking/projects/final/app/src/main/java/com/kodeco/recipefinder/data/models/SearchRecipesResponse.kt
@@ -34,8 +34,14 @@
 
 package com.kodeco.recipefinder.data.models
 
-data class Recipe(
-  val id: Int,
-  val title: String,
-  val image: String?,
+import com.squareup.moshi.Json
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class SearchRecipesResponse(
+  val offset: Int,
+  val number: Int,
+  val totalResults: Int,
+  @Json(name = "results")
+  val recipes: List<Recipe>
 )
