+++ 07-networking/projects/final/app/src/main/java/com/kodeco/recipefinder/viewmodels/RecipeViewModel.kt
@@ -38,8 +38,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.kodeco.recipefinder.data.models.Ingredient
 import com.kodeco.recipefinder.data.models.Recipe
+import com.kodeco.recipefinder.data.models.RecipeInformationResponse
 import com.kodeco.recipefinder.network.RetrofitInstance
-import com.kodeco.recipefinder.network.SpoonacularRecipe
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
@@ -73,7 +73,7 @@ class RecipeViewModel() : ViewModel() {
   private val _recipeListState = MutableStateFlow<List<Recipe>>(listOf())
   val recipeListState = _recipeListState.asStateFlow()
 
-  private val _recipeState = MutableStateFlow<SpoonacularRecipe?>(null)
+  private val _recipeState = MutableStateFlow<RecipeInformationResponse?>(null)
   val recipeState = _recipeState.asStateFlow()
 
   private val _queryState = MutableStateFlow(QueryState())
