+++ 07-networking/projects/final/gradle/libs.versions.toml
@@ -18,7 +18,7 @@ navigation="2.7.2"
 timberVersion = "5.0.1"
 kotlinx-coroutines = "1.7.2"
 retrofit="2.9.0"
-moshi="1.14.0"
+moshi="1.15.0"
 
 [libraries]
 core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "core-ktx" }
@@ -65,9 +65,9 @@ coroutines-android = { module = "org.jetbrains.kotlinx:kotlinx-coroutines-androi
 
 # Retrofit
 retrofit = {module="com.squareup.retrofit2:retrofit", version.ref="retrofit" }
-retrofit-moshi = {module="com.squareup.moshi:moshi-kotlin", version.ref="moshi" }
-retrofit-moshi-converter = {module="com.squareup.retrofit2:converter-moshi", version.ref="retrofit" }
+moshi = {module="com.squareup.moshi:moshi", version.ref="moshi" }
 moshiCodeGen = {module="com.squareup.moshi:moshi-kotlin-codegen", version.ref="moshi" }
+retrofit-moshi-converter = {module="com.squareup.retrofit2:converter-moshi", version.ref="retrofit" }
 
 [plugins]
 androidApplication = { id = "com.android.application", version.ref = "agp" }
