+++ 07-networking/projects/starter/app/src/main/java/com/kodeco/recipefinder/data/models/ExtendedIngredient.kt
@@ -32,24 +32,7 @@
  * THE SOFTWARE.
  */
 
-package com.kodeco.recipefinder.network
-
-data class SpoonacularRecipe(
-  val id: Int = 0,
-  val title: String = "",
-  val image: String? = "",
-  val imageType: String = "",
-  val summary: String = "",
-  val instructions: String = "",
-  val sourceUrl: String = "",
-
-  val preparationMinutes: Int = 0,
-  val cookingMinutes: Int = 0,
-  val extendedIngredients: List<ExtendedIngredient> = listOf(),
-  val readyInMinutes: Int = 0,
-  val servings: Int = 0,
-
-  )
+package com.kodeco.recipefinder.data.models
 
 data class ExtendedIngredient(
   val id: Int = 0,
