+++ 07-networking/projects/starter/app/src/test/java/com/kodeco/recipefinder/ExampleUnitTest.kt
@@ -34,9 +34,8 @@
 
 package com.kodeco.recipefinder
 
+import org.junit.Assert.assertEquals
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
